requirements.txt
Django==5.1.6
djangorestframework==3.15.2
environ
pyscopg2-binary
djangorestframework-simplejwt
django-extensions
django-cors-headers
PyJWT
gunicorn
